Move api method to correct class
recommended artist is in the userapi
